requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.4
python-multipart>=0.0.9
defusedxml>=0.7.1
pydantic>=2.9.0
pydantic-settings>=2.5.0
httpx>=0.27.0
pillow>=10.4.0
python-jose[cryptography]>=3.3.0
itsdangerous>=2.2.0
google-generativeai>=0.7.0
